Alleviate a race when building doc screenshots
Almost completely fixes #101 and #102.

The issue is rather simple:
 * we run an example in a virtual X display,
 * we send events to the example via xdotool,
 * immediately after, we take a screenshot (with ImageMagick).

We don't ever (and can't!) synchronize to a point "all X11 events have
been pumped, handled, and the example has redrawn" -- so there's an
obvious race happening. A little 50ms sleep should bias the race hard
enough so that we always win it.